+++ Source/Examples/ExampleLibrary/CustomSeries/MatrixSeries.cs
@@ -134,8 +134,7 @@ namespace ExampleLibrary
         /// Renders the series on the specified render context.
         /// </summary>
         /// <param name="rc">The rendering context.</param>
-        /// <param name="model">The model.</param>
-        public override void Render(IRenderContext rc, PlotModel model)
+        public override void Render(IRenderContext rc)
         {
             if (this.Matrix == null)
             {