+++ Source/OxyPlot/Axes/RangeColorAxis.cs
@@ -127,10 +127,8 @@ namespace OxyPlot.Axes
         /// Renders the axis on the specified render context.
         /// </summary>
         /// <param name="rc">The render context.</param>
-        /// <param name="model">The model.</param>
-        /// <param name="axisLayer">The rendering order.</param>
         /// <param name="pass">The render pass.</param>
-        public override void Render(IRenderContext rc, PlotModel model, AxisLayer axisLayer, int pass)
+        public override void Render(IRenderContext rc, int pass)
         {
             if (this.Position == AxisPosition.None)
             {
@@ -140,8 +138,8 @@ namespace OxyPlot.Axes
             if (pass == 0)
             {
                 double distance = this.AxisDistance;
-                double left = model.PlotArea.Left;
-                double top = model.PlotArea.Top;
+                double left = this.PlotModel.PlotArea.Left;
+                double top = this.PlotModel.PlotArea.Top;
                 double width = this.MajorTickSize - 2;
                 double height = this.MajorTickSize - 2;
 
@@ -150,20 +148,20 @@ namespace OxyPlot.Axes
                 switch (this.Position)
                 {
                     case AxisPosition.Left:
-                        left = model.PlotArea.Left - TierShift - width - distance;
-                        top = model.PlotArea.Top;
+                        left = this.PlotModel.PlotArea.Left - TierShift - width - distance;
+                        top = this.PlotModel.PlotArea.Top;
                         break;
                     case AxisPosition.Right:
-                        left = model.PlotArea.Right + TierShift + distance;
-                        top = model.PlotArea.Top;
+                        left = this.PlotModel.PlotArea.Right + TierShift + distance;
+                        top = this.PlotModel.PlotArea.Top;
                         break;
                     case AxisPosition.Top:
-                        left = model.PlotArea.Left;
-                        top = model.PlotArea.Top - TierShift - height - distance;
+                        left = this.PlotModel.PlotArea.Left;
+                        top = this.PlotModel.PlotArea.Top - TierShift - height - distance;
                         break;
                     case AxisPosition.Bottom:
-                        left = model.PlotArea.Left;
-                        top = model.PlotArea.Bottom + TierShift + distance;
+                        left = this.PlotModel.PlotArea.Left;
+                        top = this.PlotModel.PlotArea.Bottom + TierShift + distance;
                         break;
                 }
 
@@ -229,7 +227,7 @@ namespace OxyPlot.Axes
                 }
             }
 
-            var r = new HorizontalAndVerticalAxisRenderer(rc, model);
+            var r = new HorizontalAndVerticalAxisRenderer(rc, this.PlotModel);
             r.Render(this, pass);
         }