+++ Source/OxyPlot/Series/ScatterSeries{T}.cs
@@ -313,8 +313,7 @@ namespace OxyPlot.Series
         /// Renders the series on the specified rendering context.
         /// </summary>
         /// <param name="rc">The rendering context.</param>
-        /// <param name="model">The owner plot model.</param>
-        public override void Render(IRenderContext rc, PlotModel model)
+        public override void Render(IRenderContext rc)
         {
             var actualPoints = this.ActualPointsList;
             int n = actualPoints.Count;