+++ Source/OxyPlot/Series/StemSeries.cs
@@ -107,8 +107,7 @@ namespace OxyPlot.Series
         /// Renders the LineSeries on the specified rendering context.
         /// </summary>
         /// <param name="rc">The rendering context.</param>
-        /// <param name="model">The owner plot model.</param>
-        public override void Render(IRenderContext rc, PlotModel model)
+        public override void Render(IRenderContext rc)
         {
             if (this.ActualPoints.Count == 0)
             {